Fix URL in COMP25111 margin
A URL in the margin of the COMP25111 notes, which linked to a StackOverflow
post, had been setup as three text blocks. I assume this was to improve the
formatting. However, a consequence was that the notes were linking to only
the URL displayed on the first line (http://stackoverflow.com/questions).

I've fixed that at the expense of worse formatting. I did try to improve the
formatting, too, but had to get back to revision before going too far down that
rabbit hole!